Move and rename the HlslSourceHelper type
